Fm delete payment type

+++ Bangazon/Controllers/PaymentTypesController.cs
@@ -161,10 +161,21 @@ namespace Bangazon.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var paymentType = await _context.PaymentType.FindAsync(id);
-            _context.PaymentType.Remove(paymentType);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var paymentAssigned = await _context.Order.FirstOrDefaultAsync(o => o.PaymentTypeId == id);
+
+            if( paymentAssigned != null)
+            {
+               TempData["ErrorMessage"] = "You can't delete this paymeny type because it has been assigned to an existing order";
+               return RedirectToAction(nameof(Index));
+            }
+            else 
+            {
+                var paymentType = await _context.PaymentType.FindAsync(id);
+                _context.PaymentType.Remove(paymentType);
+                await _context.SaveChangesAsync();
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
 
         private bool PaymentTypeExists(int id)
@@ -173,5 +184,7 @@ namespace Bangazon.Controllers
         }
 
         private Task<ApplicationUser> GetCurrentUserAsync() => _userManager.GetUserAsync(HttpContext.User);
+
+        
     }
 }
